python_rag_server/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
chromadb==0.4.18
sentence-transformers==2.2.2
python-multipart==0.0.6
python-docx==1.1.0
PyPDF2==3.0.1
openpyxl==3.1.2
pandas==2.1.4
numpy==1.24.3
torch==2.1.1
transformers==4.36.2
pydantic==2.5.2
httpx==0.25.2
aiofiles==23.2.0
python-magic==0.4.27
chardet==5.2.0
tiktoken==0.5.2
